third_party/blink/web_tests/fast/forms/form-radio-node-list-expected.txt

Test RadioNodeLists returned by the HTMLFormElement named-getter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS form1.elements.length is 9
Check that only 'listed elements' are included in the list, if any.
PASS radioNodeList.length is 9
PASS radioNodeList[0] instanceof HTMLButtonElement is true
PASS radioNodeList[1] instanceof HTMLFieldSetElement is true
PASS radioNodeList[2] instanceof HTMLInputElement is true
PASS radioNodeList[2].type is "hidden"
PASS radioNodeList[3] instanceof HTMLInputElement is true
PASS radioNodeList[3].type is "text"
PASS radioNodeList[4] instanceof HTMLInputElement is true
PASS radioNodeList[4].type is "radio"
PASS radioNodeList[5] instanceof HTMLOutputElement is true
PASS radioNodeList[6] instanceof HTMLObjectElement is true
PASS radioNodeList[7] instanceof HTMLSelectElement is true
PASS radioNodeList[8] instanceof HTMLTextAreaElement is true
PASS successfullyParsed is true

TEST COMPLETE

    
 
